pyproject.toml
[project]
name = "pn-ai-foundry-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "azure-ai-projects (>=1.0.0b12,<2.0.0)",
    "azure-identity (>=1.23.0,<2.0.0)",
    "azure-ai-inference (>=1.0.0b9,<2.0.0)",
    "azure-search-documents (>=11.5.3,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
